Tested Model

+++ src/downloadGames.py
@@ -1,16 +1,24 @@
 from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
 from bs4 import BeautifulSoup
 import requests
 import time
+import os
+
+existing_games = os.listdir("data/games/")
+print(existing_games)
 
 for p in range(1,30):
-    driver = webdriver.Chrome()
+    options = Options()
+    options.add_argument("--headless=new") 
+    driver = webdriver.Chrome(options=options)
+    options.add_argument("--headless=new")
     driver.get(f"https://replay.pokemonshowdown.com/?format=%5BGen%209%5D%20VGC%202025%20Reg%20H%20(Bo3)&page={p}&sort=rating")
 
-    time.sleep(2)
+    time.sleep(1)
+
     html = driver.page_source
     soup = BeautifulSoup(html, "html.parser")
-    print(soup.prettify)
 
     links = []
 
@@ -21,10 +29,12 @@ for p in range(1,30):
         links.append(a.get("href"))
 
     for link in links:
-        url = f"https://replay.pokemonshowdown.com/{link}.log"
-        page = requests.get(url)
-        soup = BeautifulSoup(page.content, "html.parser")
-        with open(f"data/games/{link}", 'w', encoding="utf-8") as file:
-            file.write(soup.text)
+        if link not in existing_games:
+            url = f"https://replay.pokemonshowdown.com/{link}.log"
+            page = requests.get(url)
+            soup = BeautifulSoup(page.content, "html.parser")
+            print(link)
+            with open(f"data/games/{link}", 'w', encoding="utf-8") as file:
+                file.write(soup.text)
